+++ docs/examples/examples_index.rst
@@ -2,6 +2,7 @@
    Interactive experiment run comparison <interactive_compare_experiments>
    Hyper parameter comparison <explore_hyperparams_skopt>
    Run skopt/hyperopt hyperparameter sweep <run_hyperparameter_search>
+   Log model diagnostics <log_model_diagnostics>
    Monitor lightGBM training <monitor_lgbm>
    Monitor fast.ai training <monitor_fastai>
    Log matplotlib charts to neptune <log_matplotlib>
