--- docs/user_guide/viz/utils.rst
@@ -1,6 +1,0 @@
-Visualization utils
-===========
-
-.. automodule:: neptunecontrib.viz.utils
-    :members:
-    :show-inheritance:
